Implement profile update
This implements the profile update method in the minder server.

This was done last as it's kinda complex... as you can tell from the PR.

We have to verify that the new profile is valid, that certain values like the
project, name and provider don't change. Then we update the rules, and clean up
the unused ones. And keep the instantiations in check.

This, however, will be a very nice usability improvement.

This also adds a profile update command